TT-3021 use edited workflow
- handle name change better
- move edit button to team screen
